Fully removed CBitcoinAddress(except for tests), fixed a couple of warnings
ScriptSigArgsExpected is removed because it is not used

+++ src/darksend.cpp
@@ -947,7 +947,7 @@ bool CDarkSendPool::SignatureValid(const CScript& newSig, const CTxIn& newVin){
         CScriptWitness *witness;
         int flags = SCRIPT_VERIFY_P2SH | SCRIPT_VERIFY_STRICTENC;
         //If transaction contains witness, then witness script should be verified
-        if (n < txNew.wit.vtxinwit.size()) {
+        if (n < (int)txNew.wit.vtxinwit.size()) {
             witness = &txNew.wit.vtxinwit[n].scriptWitness;
             flags |= SCRIPT_VERIFY_WITNESS;
         } else {
